Add files via upload
insert_venta.py was modified to contain a confirmation and injection to "Deudores" table in which I have a record of my clients' debts.
update_venta.py is added, which is an updater of those entries in "Ventas" table whose second date is None, i.e. there is an active debt. Then by running this script the table updates with a new payment date and amount.